Move ProblemFiltersConfig to functional-tests
